[003-IMAGE] refactoring package
